face_recognition_system-main/backend/requirements.txt
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.1
annotated-types==0.7.0
bcrypt==4.0.1
blinker==1.9.0
boto3==1.38.27
botocore==1.38.27
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
dlib==20.0.0
easydict==1.13
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flatbuffers==25.2.10
fonttools==4.59.0
greenlet==3.2.2
gunicorn==21.2.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy>=1.26.0,<2.0.0
onnx==1.18.0
onnxruntime==1.22.1
# Chỉ dùng một trong hai opencv packages
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas>=2.2.3
pillow==11.3.0
pluggy==1.6.0
prettytable==3.16.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.0.7
requests==2.32.4
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
simsimd==6.5.0
six==1.17.0
spicy==0.16.0
SQLAlchemy==2.0.41
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
XlsxWriter==3.1.9